requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-multipart
websockets
itsdangerous
bcrypt==4.0.1
python-dotenv
python-jose[cryptography]
jinja2
python-socketio
google-generativeai
